backend/requirements.txt
sqlalchemy==1.4.41
psycopg2-binary==2.9.5
psycopg==3.1.12
python-dotenv==1.0.0
flask==2.2.5
flask-sqlalchemy==3.0.5
flask-cors==3.0.10
requests==2.28.2
beautifulsoup4==4.11.2
lxml==4.9.2
html5lib==1.1
gunicorn==20.1.0
urllib3==1.26.15 